requirements.txt
jinja2
aiohttp                                                                                                                                                                                                                              │
fastapi                                                                                                                                                                                                                              │
uvicorn                                                                                                                                                                                                                              │
python-dotenv                                                                                                                                                                                                                        │
pymongo                                                                                                                                                                                                                              │
pydantic                                                                                                                                                                                                                             │
langchain                                                                                                                                                                                                                            │
langgraph                                                                                                                                                                                                                            │
langchain_openai                                                                                                                                                                                                                     │
langchain_core    